Add some text

+++ Example004_Div/Program.cs
@@ -1,7 +1,10 @@
 ﻿Console.Clear();
 Console.Write("Введите первое число: ");
-int NumberA = Convert.ToInt32(Console.ReadLine());
+double NumberA = Convert.ToInt32(Console.ReadLine()); // int для целых чисел, double для вещественных(со знаками после запятой)
 Console.Write("Введите второе число: ");
-int NumberB = Convert.ToInt32(Console.ReadLine());
-int Sum = NumberA / NumberB;
+double NumberB = Convert.ToInt32(Console.ReadLine());
+double Sum = NumberA / NumberB;
 Console.WriteLine(Sum);
+
+// значок % означает нахождение остатка от деления
+// new Random().Next(min, max); даст случайное целое число от min до max1
